requirements.txt
streamlit
moviepy
pillow
opencv-python-headless
numpy